moved and implemented spaceship operator
