--- include/scouting-iterator.h
@@ -1,282 +1,0 @@
-#ifndef SCOUTING_ITERATOR
-#define SCOUTING_ITERATOR
-
-#include <iterator>    // type traits
-
-
-// A scouting iterator is a pair of iterators designed for 2-stage traversal, where one iterator progresses while the
-//     other maintains a previous position.
-
-// Based on the design of __normal_iterator from GCC 9 (stl_iterator.h)
-// Container is used to distinguish iterators of containers with the same underlying type
-template <typename Iter, typename Container>
-class scouting_iterator
-{
-    using traits_type = std::iterator_traits<Iter>;
-
-    template<typename _Iter>
-    using convertible_type = scouting_iterator<
-        _Iter,
-        typename std::enable_if<
-            std::is_same_v<_Iter, typename Container::pointer>,
-            Container
-            >::type
-        >;
-
-    using self_type = scouting_iterator<Iter, Container>;
-
-    Iter sentry = Iter {};
-    Iter scout  = Iter {};
-
-public:
-    // Traits
-    using iterator_category = typename traits_type::iterator_category;
-    using value_type        = typename traits_type::value_type;
-    using pointer           = typename traits_type::pointer;
-    using reference         = typename traits_type::reference;
-    using difference_type   = typename traits_type::difference_type;
-    using size_type         = std::size_t;
-
-
-    // Constructors
-    constexpr scouting_iterator () noexcept
-        : sentry {Iter()}, scout {sentry}
-    {}
-
-    explicit constexpr scouting_iterator (const value_type* i) noexcept
-        : sentry {i}, scout {sentry}
-    {}
-
-    explicit constexpr scouting_iterator (const Iter& sentry, const Iter& scout) noexcept
-        : sentry {sentry}, scout {scout}
-    {}
-
-    // Allow iterator to const_iterator conversion
-    template <typename _Iter>
-    constexpr scouting_iterator (const convertible_type<_Iter>& i) noexcept
-        : sentry {i.get_sentry()}, scout {sentry}
-    {}
-
-    template <typename _Iter>
-    self_type& operator= (const convertible_type<_Iter>& i) noexcept
-    {
-        scout  = i.scout;
-        sentry = i.sentry;
-        return *this;
-    }
-
-    self_type& operator= (const Iter& i) noexcept
-    {
-        scout = i;
-        return *this;
-    }
-
-
-    // Element access
-    self_type& save () noexcept
-    {
-        sentry = scout;
-        return *this;
-    }
-
-    constexpr self_type& restore () noexcept
-    {
-        scout = sentry;
-        return *this;
-    }
-
-    constexpr Iter  get_sentry  () const noexcept    { return sentry; }
-    constexpr Iter  get_scout   () const noexcept    { return scout;  }
-    constexpr Iter& sentry_base ()       noexcept    { return sentry; }
-    constexpr Iter& base        ()       noexcept    { return scout;  }
-    constexpr const value_type* saved_data () const noexcept    { return &*sentry; }
-    constexpr const value_type* data       () const noexcept    { return &*scout; }
-
-
-    // Operations
-    reference operator*  () const noexcept    { return *scout; }
-    pointer   operator-> () const noexcept    { return scout;  }
-
-    self_type& operator++ () noexcept
-    {
-        ++scout;
-        return *this;
-    }
-
-    self_type operator++ (int) noexcept
-    {
-        return {sentry, scout++};
-    }
-
-    self_type& operator-- () noexcept
-    {
-        --scout;
-        return *this;
-    }
-
-    self_type operator-- (int) noexcept
-    {
-        return {sentry, scout--};
-    }
-
-    reference operator[] (difference_type n) const noexcept
-    {
-        return scout[n];
-    }
-
-    self_type& operator+= (difference_type n) noexcept
-    {
-        scout += n;
-        return *this;
-    }
-
-    self_type operator+ (difference_type n) const noexcept
-    {
-        return {sentry, scout + n};
-    }
-
-    friend self_type operator+ (difference_type n, const self_type& i) noexcept
-    {
-        return {i.sentry, i.scout + n};
-    }
-
-    self_type& operator-= (difference_type n) noexcept
-    {
-        scout -= n;
-        return *this;
-    }
-
-    self_type operator- (difference_type n)
-    {
-        return {sentry, scout - n};
-    }
-
-    difference_type operator- (const Iter& i) const noexcept
-    {
-        return scout - i;
-    }
-
-    friend difference_type operator- (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs - rhs.scout;
-    }
-
-    bool operator== (const Iter& i) const noexcept
-    {
-        return scout == i;
-    }
-
-    friend bool operator== (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs == rhs.scout;
-    }
-
-    bool operator!= (const Iter& i) const noexcept
-    {
-        return scout != i;
-    }
-
-    friend bool operator!= (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs != rhs.scout;
-    }
-
-    bool operator< (const Iter& i) const noexcept
-    {
-        return scout < i;
-    }
-
-    friend bool operator< (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs < rhs.scout;
-    }
-
-    bool operator<= (const Iter& i) const noexcept
-    {
-        return scout <= i;
-    }
-
-    friend bool operator<= (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs <= rhs.scout;
-    }
-
-    bool operator> (const Iter& i) const noexcept
-    {
-        return scout > i;
-    }
-
-    friend bool operator> (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs > rhs.scout;
-    }
-
-    bool operator>= (const Iter& i) const noexcept
-    {
-        return scout >= i;
-    }
-
-    friend bool operator>= (const Iter& lhs, const self_type& rhs) noexcept
-    {
-        return lhs >= rhs.scout;
-    }
-
-    constexpr difference_type distance () const noexcept
-    {
-        return scout - sentry;
-    }
-};
-
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator== (const scouting_iterator<IterL, Container>& lhs,
-                        const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() == rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator!= (const scouting_iterator<IterL, Container>& lhs,
-                        const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() != rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator< (const scouting_iterator<IterL, Container>& lhs,
-                       const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() < rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator<= (const scouting_iterator<IterL, Container>& lhs,
-                        const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() <= rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator> (const scouting_iterator<IterL, Container>& lhs,
-                       const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() > rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline bool operator>= (const scouting_iterator<IterL, Container>& lhs,
-                        const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() >= rhs.get_scout();
-}
-
-template <typename IterL, typename IterR, typename Container>
-inline typename scouting_iterator<IterL, Container>::difference_type
-operator- (const scouting_iterator<IterL, Container>& lhs,
-           const scouting_iterator<IterR, Container>& rhs) noexcept
-{
-    return lhs.get_scout() - rhs.get_scout();
-}
-
-
-#endif // SCOUTING_ITERATOR
